other_projects/ai_vids_virality/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
pyyaml>=6.0
pytest>=8.0
httpx>=0.27
pillow>=10.0
requests>=2.31
feedparser>=6.0
yt-dlp>=2024.0
python-dotenv>=1.0
